error response for delete api

+++ student/src/main/java/vn/edu/vnua/fita/student/service/admin/management/StatusManager.java
@@ -1,6 +1,7 @@
 package vn.edu.vnua.fita.student.service.admin.management;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -17,7 +18,8 @@ public class StatusManager implements IStatusService {
     private final StatusRepository statusRepository;
     private final String statusHadExisted = "Mã trạng thái đã tồn tại trong hệ thống";
     private final String statusNotFound = "Mã trạng thái không tồn tại trong hệ thống";
-    private final String cannotDelete = "Không thể xoá trạng thái này";
+    private final String cannotDeleteDefault = "Không thể xoá trạng thái mặc định, gồm: 'Đã nhập học', 'Đã bỏ học, 'Đã xin thôi học', 'Bị buộc thôi học' và 'Đã tốt nghiệp'";
+    private final String cannotDeleteAssigned = "Trạng thái này đang được gán cho sinh viên, vui lòng xoá hết các sinh viên có trạng thái này trước khi xoá trạng thái. Nếu không, có thể tiến hành cập nhật trạng thái thay vì xoá";
 
     @Override
     public Page<Status> getStatusList(GetStatusListRequest request) {
@@ -41,11 +43,16 @@ public class StatusManager implements IStatusService {
 
     @Override
     public Status deleteStatus(Integer id) {
-        if (id == 1 || id == 2 || id == 3) {
-            throw new RuntimeException(cannotDelete);
+        try {
+            if (id == 1 || id == 2 || id == 3 || id == 4 || id == 5) {
+                throw new RuntimeException(cannotDeleteDefault);
+            }
+            Status status = statusRepository.findById(id).orElseThrow(() -> new RuntimeException(statusNotFound));
+            statusRepository.deleteById(id);
+            return status;
+        } catch (
+                DataIntegrityViolationException e) {
+            throw new RuntimeException(cannotDeleteAssigned);
         }
-        Status status = statusRepository.findById(id).orElseThrow(() -> new RuntimeException(statusNotFound));
-        statusRepository.deleteById(id);
-        return status;
     }
 }
